close to being finished
Close to being fully polished, minimax funcitno works as planned, sometimes it makes very strange choices, but those choices still result in a win or tie later. It will play a move to stop you from winning instead of winning, but the move that stops you from winning, creates a circumstance where it is impossible for it to lose. Perhaps it evaluates that both moves are winning and plays the first move that is winning.